refactor: decompose Viewer.tsx into modular components and enhance reliability
